requirements.txt
# python 3.10.10
Flask
gradio==4.25.0
jieba
numpy
onnxruntime
opencv_contrib_python
opencv_python
opencv_python_headless
Pillow
fastapi
